+++ frontend/src/Tipo-Producto.jsx
@@ -7,47 +7,55 @@ import { Menu } from "./Menu";
 
 export function TipoProducto(){
     const [tipo_producto, setTipoProducto] = useState([]);
-    const [mensaje, setMensaje] = useState('');
+    const [producto, setProducto] = useState([]);
 
     useEffect(() => {
         API.getTipoProductos().then(setTipoProducto);
+        API.getProductos().then(setProducto);
     }, []);
 
     const eliminar = (e, id_tipo_producto) => {
         e.preventDefault();
-
-        if (id_tipo_producto === 1 || id_tipo_producto===2 || id_tipo_producto===3) {
+        if (id_tipo_producto === 1 || id_tipo_producto === 2 || id_tipo_producto === 3) {
             Swal.fire({
                 icon: 'error',
                 title: 'No permitido',
                 text: 'No puedes eliminar los Elementos deportivos, Suplementos y Pelotas',
             });
         } else {
+            // Verificar si el tipo de producto está en uso en productos
+            const tipoProductoEnUso = producto.some(producto => producto.id_tipo_producto === id_tipo_producto);
     
-
-        Swal.fire({
-            title: '¿Estás seguro?',
-            text: "¡No podrás revertir esto!",
-            icon: 'warning',
-            showCancelButton: true,
-            confirmButtonColor: '#d33',
-            cancelButtonColor: '#3085d6',
-            confirmButtonText: '¡Sí, bórralo!'
-        }).then((result) => {
-            if (result.isConfirmed) {
-               const respuesta= API.deleteTipoProducto(id_tipo_producto);
-                API.getTipoProductos().then(setTipoProducto);
-                Swal.fire(
-                    '¡Eliminado!',
-                    'El tipo de producto ha sido eliminado.',
-                    'success'
-                    
-                );
-                
+            if (tipoProductoEnUso) {
+                Swal.fire({
+                    icon: 'error',
+                    title: 'No permitido',
+                    text: 'No puedes eliminar este tipo de producto porque está en uso en Productos.',
+                });
+            } else {
+                Swal.fire({
+                    title: '¿Estás seguro?',
+                    text: '¡No podrás revertir esto!',
+                    icon: 'warning',
+                    showCancelButton: true,
+                    confirmButtonColor: '#d33',
+                    cancelButtonColor: '#3085d6',
+                    confirmButtonText: '¡Sí, bórralo!'
+                }).then((result) => {
+                    if (result.isConfirmed) {
+                        API.deleteTipoProducto(id_tipo_producto);
+                        API.getTipoProductos().then(setTipoProducto);
+                        Swal.fire(
+                            '¡Eliminado!',
+                            'El tipo de producto ha sido eliminado.',
+                            'success'
+                        );
+                    }
+                });
             }
-        });
-    };}
-
+        }
+    };
+    
     return (
         <>
            <Menu/>
